Step 1 - Importing Data
Just imported the data into SQL to perform operations. This was a CSV file which is not very big not small too i want to test it by converting it myself manually to SQL and perform operations on it. I have the problem statement and some questions for this data and based on them i am going to work on the dataset. Will add everything related in readme.